Refactor UML diagram generation UI to a unified view and retain source code
- Created `UmlDiagramUIFactory` to centralize UI components.
- Refactored `FlowDiagramAction`, `SequenceDiagramAction`, and `ClassDiagramAction` to utilize the common factory.
- Introduced a `JTabbedPane` layout with 'Source Code View' and 'PlantUML Code View' to prevent overwriting user input and parsed code when AI generation completes.
