+++ src/main/java/com/huq/idea/flow/apidoc/ClassDiagramAction.java
@@ -32,11 +32,10 @@ import com.intellij.psi.PsiType;
 import com.intellij.psi.util.PsiTreeUtil;
 import com.intellij.psi.util.PsiUtil;
 import org.jetbrains.annotations.NotNull;
+import com.huq.idea.flow.apidoc.ui.UmlDiagramUIFactory;
 
 import javax.swing.*;
 import java.awt.*;
-import java.awt.datatransfer.Clipboard;
-import java.awt.datatransfer.StringSelection;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -123,8 +122,16 @@ public class ClassDiagramAction extends AnAction {
                 String collectedCode = ReadAction.compute(() -> collectCodeFromClasses(associatedClasses));
 
                 // 显示初始对话框
-                SwingUtilities.invokeLater(() ->
-                    showInitialDialog(project, collectedCode, currentClass.getName()));
+                SwingUtilities.invokeLater(() -> {
+                    UmlDiagramUIFactory.PromptProvider promptProvider = new UmlDiagramUIFactory.PromptProvider() {
+                        @Override
+                        public String getPrompt(String collectedCode) {
+                            String classDiagramPromptTemplate = IdeaSettings.getInstance().getState().getClassDiagramPrompt();
+                            return String.format(classDiagramPromptTemplate, collectedCode);
+                        }
+                    };
+                    UmlDiagramUIFactory.showInitialDialog(project, collectedCode, "UML类图: " + currentClass.getName(), promptProvider, "生成类图");
+                });
             }
         }.queue();
     }
@@ -248,336 +255,5 @@ public class ClassDiagramAction extends AnAction {
             }
         }
         return codeBuilder.toString();
-    }
-
-    private void showInitialDialog(Project project, String collectedCode, String title) {
-        JTabbedPane tabbedPane = new JTabbedPane();
-        JPanel plantUmlTab = createInitialPlantUmlTab(project, title, collectedCode);
-        tabbedPane.addTab("PlantUML视图", plantUmlTab);
-
-        JPanel bottomPanel = new JPanel(new FlowLayout(FlowLayout.RIGHT));
-        JLabel enhancedLabel = new JLabel();
-        enhancedLabel.setForeground(Color.BLUE);
-        bottomPanel.add(enhancedLabel);
-
-        JPanel mainPanel = new JPanel(new BorderLayout());
-        mainPanel.add(tabbedPane, BorderLayout.CENTER);
-        mainPanel.add(bottomPanel, BorderLayout.SOUTH);
-
-        UmlFlowService plugin = project.getService(UmlFlowService.class);
-        mainPanel.setName(title);
-        plugin.addFlow(mainPanel);
-    }
-
-    private JPanel createInitialPlantUmlTab(Project project, String title, String collectedCode) {
-        JPanel panel = new JPanel(new BorderLayout());
-
-        JSplitPane splitPane = new JSplitPane(JSplitPane.HORIZONTAL_SPLIT);
-        splitPane.setDividerLocation(500);
-        splitPane.setResizeWeight(0.5);
-
-        JPanel codePanel = new JPanel(new BorderLayout());
-        JTextArea textArea = new JTextArea();
-        textArea.setEditable(true);
-        textArea.setText(collectedCode);
-        textArea.setFont(new Font(Font.MONOSPACED, Font.PLAIN, 14));
-        codePanel.add(new JScrollPane(textArea), BorderLayout.CENTER);
-
-        JPanel diagramPanel = new JPanel(new BorderLayout());
-        JLabel waitingLabel = new JLabel("点击\"生成类图\"按钮生成类图", SwingConstants.CENTER);
-        waitingLabel.setFont(new Font(Font.SANS_SERIF, Font.BOLD, 16));
-        diagramPanel.add(waitingLabel, BorderLayout.CENTER);
-
-        splitPane.setLeftComponent(codePanel);
-        splitPane.setRightComponent(diagramPanel);
-
-        JPanel buttonPanel = new JPanel(new FlowLayout(FlowLayout.RIGHT));
-
-        JLabel providerLabel = new JLabel("AI提供商:");
-        java.util.List<IdeaSettings.CustomAiProviderConfig> availableProviders = AiUtils.getCustomProviders();
-        JComboBox<IdeaSettings.CustomAiProviderConfig> providerComboBox = new JComboBox<>(availableProviders.toArray(new IdeaSettings.CustomAiProviderConfig[0]));
-
-        JLabel specificModelLabel = new JLabel("具体模型:");
-        JComboBox<String> specificModelComboBox = new JComboBox<>();
-
-        providerComboBox.addActionListener(e -> {
-            IdeaSettings.CustomAiProviderConfig selected = (IdeaSettings.CustomAiProviderConfig) providerComboBox.getSelectedItem();
-            specificModelComboBox.removeAllItems();
-            if (selected != null && selected.getModels() != null) {
-                String[] models = selected.getModels().split(",");
-                for (String model : models) {
-                    specificModelComboBox.addItem(model.trim());
-                }
-                if (models.length > 0) {
-                    specificModelComboBox.setSelectedIndex(0);
-                }
-            }
-        });
-
-        if (!availableProviders.isEmpty()) {
-            providerComboBox.setSelectedIndex(-1);
-            providerComboBox.setSelectedIndex(0);
-        }
-
-        providerComboBox.setRenderer(new DefaultListCellRenderer() {
-            @Override
-            public Component getListCellRendererComponent(JList<?> list, Object value, int index, boolean isSelected, boolean cellHasFocus) {
-                super.getListCellRendererComponent(list, value, index, isSelected, cellHasFocus);
-                if (value instanceof IdeaSettings.CustomAiProviderConfig) {
-                    IdeaSettings.CustomAiProviderConfig provider = (IdeaSettings.CustomAiProviderConfig) value;
-                    setText(provider.getName());
-                }
-                return this;
-            }
-        });
-
-        buttonPanel.add(providerLabel);
-        buttonPanel.add(providerComboBox);
-        buttonPanel.add(specificModelLabel);
-        buttonPanel.add(specificModelComboBox);
-
-        JButton generateButton = new JButton("生成类图");
-        generateButton.addActionListener(e -> {
-            generateButton.setEnabled(false);
-            generateButton.setText("生成中...");
-
-            IdeaSettings.CustomAiProviderConfig selectedProvider = (IdeaSettings.CustomAiProviderConfig) providerComboBox.getSelectedItem();
-            String selectedModel = (String) specificModelComboBox.getSelectedItem();
-
-            if (selectedProvider == null) {
-                Notifications.Bus.notify(new Notification(
-                        "com.yt.huq.idea",
-                        "无可用AI模型",
-                        "未选择AI提供商或没有可用的AI提供商。请在设置 > UmlFlowAiConfigurable 中配置至少一个提供商",
-                        NotificationType.ERROR),
-                        project);
-                generateButton.setEnabled(true);
-                generateButton.setText("生成类图");
-                return;
-            }
-            if (IdeaSettings.getInstance().getState().getPlantumlPathVal() == null ||
-                    IdeaSettings.getInstance().getState().getPlantumlPathVal().trim().isEmpty()) {
-                Notifications.Bus.notify(new Notification(
-                        "com.yt.huq.idea",
-                        "PlantUML路径缺失",
-                        "PlantUML路径未配置。请在设置 > UmlFlowAiConfigurable 中设置",
-                        NotificationType.ERROR),
-                        project);
-                generateButton.setEnabled(true);
-                generateButton.setText("生成类图");
-                return;
-            }
-
-            new Task.Backgroundable(project, "生成类图", true, PerformInBackgroundOption.ALWAYS_BACKGROUND) {
-                @Override
-                public void run(@NotNull ProgressIndicator indicator) {
-                    indicator.setText("正在生成PlantUML类图...");
-                    indicator.setIndeterminate(true);
-
-                    String classDiagramPromptTemplate = IdeaSettings.getInstance().getState().getClassDiagramPrompt();
-                    String flowPrompt = String.format(classDiagramPromptTemplate, collectedCode);
-
-                    AiUtils.AiConfig config = new AiUtils.AiConfig(selectedProvider, selectedModel);
-                    if (config.getApiKey() == null || config.getApiKey().trim().isEmpty()) {
-                        SwingUtilities.invokeLater(() -> {
-                            generateButton.setEnabled(true);
-                            Notifications.Bus.notify(new Notification(
-                                "com.yt.huq.idea",
-                                "API密钥未配置",
-                                "请在设置中为 " + selectedProvider.getName() + " 配置API密钥",
-                                NotificationType.WARNING),
-                                project);
-                        });
-                        return;
-                    }
-
-                    config.setSystemMessage("你是一个专业的PlantUML类图生成专家，擅长分析Java代码并生成高质量的类图。")
-                          .setTemperature(0.7)
-                          .setMaxTokens(8000);
-
-                    AiUtils.AiResponse response = AiUtils.callAi(flowPrompt, config);
-                    String classDiagram = response.isSuccess() ? response.getContent() : null;
-
-                    if (classDiagram != null && !classDiagram.isEmpty()) {
-                        classDiagram = cleanupUmlResponse(classDiagram);
-
-                        String finalClassDiagram = classDiagram;
-                        SwingUtilities.invokeLater(() -> {
-                            textArea.setText(finalClassDiagram);
-                            JPanel newDiagramPanel = PlantUmlRenderer.createPlantUmlPanel(finalClassDiagram);
-                            splitPane.setRightComponent(newDiagramPanel);
-
-                            generateButton.setEnabled(true);
-                            generateButton.setText("重新生成");
-
-                            panel.revalidate();
-                            panel.repaint();
-                        });
-                    } else {
-                        SwingUtilities.invokeLater(() -> {
-                            String errorMsg = "生成类图失败，请检查API设置和网络连接";
-                            if (response != null && !response.isSuccess() && response.getErrorMessage() != null) {
-                                errorMsg = "生成类图失败: " + response.getErrorMessage();
-                            }
-
-                            Notifications.Bus.notify(new Notification(
-                                    "com.yt.huq.idea",
-                                    "类图生成",
-                                    errorMsg,
-                                    NotificationType.ERROR),
-                                    project);
-
-                            generateButton.setEnabled(true);
-                            generateButton.setText("生成类图");
-                        });
-                    }
-                }
-            }.queue();
-        });
-        buttonPanel.add(generateButton);
-
-        JButton copyButton = new JButton("复制到剪贴板");
-        copyButton.addActionListener(e -> {
-            copyToClipboard(textArea.getText());
-            Notifications.Bus.notify(new Notification(
-                    "com.yt.huq.idea",
-                    "UML图表",
-                    "UML图表已复制到剪贴板",
-                    NotificationType.INFORMATION),
-                    project);
-        });
-        buttonPanel.add(copyButton);
-
-        JButton saveCodeButton = new JButton("保存代码");
-        saveCodeButton.addActionListener(e -> {
-            com.intellij.openapi.fileChooser.FileChooserDescriptor descriptor =
-                com.intellij.openapi.fileChooser.FileChooserDescriptorFactory.createSingleFolderDescriptor();
-            descriptor.setTitle("选择保存目录");
-
-            com.intellij.openapi.vfs.VirtualFile dir = com.intellij.openapi.fileChooser.FileChooser.chooseFile(descriptor, project, null);
-            if (dir != null) {
-                java.io.File fileToSave = new java.io.File(dir.getPath(), title.replaceAll("[^a-zA-Z0-9]", "_") + ".puml");
-                try {
-                    java.io.FileWriter writer = new java.io.FileWriter(fileToSave);
-                    writer.write(textArea.getText());
-                    writer.close();
-                    Notifications.Bus.notify(new Notification(
-                            "com.yt.huq.idea",
-                            "UML图表",
-                            "UML代码已保存到 " + fileToSave.getAbsolutePath(),
-                            NotificationType.INFORMATION),
-                            project);
-                } catch (Exception ex) {
-                    Notifications.Bus.notify(new Notification(
-                            "com.yt.huq.idea",
-                            "UML图表",
-                            "保存UML代码失败: " + ex.getMessage(),
-                            NotificationType.ERROR),
-                            project);
-                }
-            }
-        });
-        buttonPanel.add(saveCodeButton);
-
-        JButton saveImageButton = new JButton("保存图像");
-        saveImageButton.addActionListener(e -> {
-            if (textArea.getText().trim().isEmpty() || !textArea.getText().contains("@startuml")) {
-                Notifications.Bus.notify(new Notification(
-                        "com.yt.huq.idea",
-                        "UML图表",
-                        "没有有效的UML图表可以保存",
-                        NotificationType.WARNING),
-                        project);
-                return;
-            }
-
-            com.intellij.openapi.fileChooser.FileChooserDescriptor descriptor =
-                com.intellij.openapi.fileChooser.FileChooserDescriptorFactory.createSingleFolderDescriptor();
-            descriptor.setTitle("选择保存目录");
-
-            com.intellij.openapi.vfs.VirtualFile dir = com.intellij.openapi.fileChooser.FileChooser.chooseFile(descriptor, project, null);
-            if (dir != null) {
-                java.io.File fileToSave = new java.io.File(dir.getPath(), title.replaceAll("[^a-zA-Z0-9]", "_") + ".png");
-                try {
-                    byte[] pngData = PlantUmlRenderer.renderPlantUmlToPng(textArea.getText());
-                    if (pngData != null) {
-                        java.io.FileOutputStream fos = new java.io.FileOutputStream(fileToSave);
-                        fos.write(pngData);
-                        fos.close();
-                        Notifications.Bus.notify(new Notification(
-                                "com.yt.huq.idea",
-                                "UML图表",
-                                "UML图像已保存到 " + fileToSave.getAbsolutePath(),
-                                NotificationType.INFORMATION),
-                                project);
-                    }
-                } catch (PlantUmlRenderException ex) {
-                    LOG.error("Failed to render UML diagram", ex);
-                    Notifications.Bus.notify(new Notification(
-                            "com.yt.huq.idea",
-                            "UML图表",
-                            "渲染UML图像失败: " + ex.getMessage(),
-                            NotificationType.ERROR),
-                            project);
-                } catch (Exception ex) {
-                    Notifications.Bus.notify(new Notification(
-                            "com.yt.huq.idea",
-                            "UML图表",
-                            "保存UML图像失败: " + ex.getMessage(),
-                            NotificationType.ERROR),
-                            project);
-                }
-            }
-        });
-        buttonPanel.add(saveImageButton);
-
-        JButton refreshButton = new JButton("刷新图像");
-        refreshButton.addActionListener(e -> {
-            String updatedUmlContent = textArea.getText();
-            JPanel newDiagramPanel = PlantUmlRenderer.createPlantUmlPanel(updatedUmlContent);
-            splitPane.setRightComponent(newDiagramPanel);
-            splitPane.revalidate();
-            splitPane.repaint();
-            panel.revalidate();
-            panel.repaint();
-        });
-        buttonPanel.add(refreshButton);
-
-        panel.add(splitPane, BorderLayout.CENTER);
-        panel.add(buttonPanel, BorderLayout.SOUTH);
-
-        return panel;
-    }
-
-    private void copyToClipboard(String text) {
-        StringSelection stringSelection = new StringSelection(text);
-        Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
-        clipboard.setContents(stringSelection, null);
-    }
-
-    private String cleanupUmlResponse(String umlResponse) {
-        if (umlResponse == null || umlResponse.isEmpty()) {
-            return umlResponse;
-        }
-
-        if (!umlResponse.startsWith("@startuml")) {
-            int startIndex = umlResponse.indexOf("@startuml");
-            if (startIndex >= 0) {
-                umlResponse = umlResponse.substring(startIndex);
-            } else {
-                umlResponse = "@startuml\n" + umlResponse;
-            }
-        }
-
-        if (!umlResponse.endsWith("@enduml")) {
-            if (umlResponse.contains("@enduml")) {
-                int endIndex = umlResponse.lastIndexOf("@enduml") + "@enduml".length();
-                umlResponse = umlResponse.substring(0, endIndex);
-            } else {
-                umlResponse = umlResponse + "\n@enduml";
-            }
-        }
-
-        return umlResponse;
     }
 }
