+++ src/main/java/com/huq/idea/flow/apidoc/SequenceDiagramAction.java
@@ -27,11 +27,10 @@ import com.intellij.psi.PsiFile;
 import com.intellij.psi.PsiJavaFile;
 import com.intellij.psi.PsiMethod;
 import org.jetbrains.annotations.NotNull;
+import com.huq.idea.flow.apidoc.ui.UmlDiagramUIFactory;
 
 import javax.swing.*;
 import java.awt.*;
-import java.awt.datatransfer.Clipboard;
-import java.awt.datatransfer.StringSelection;
 
 /**
  * Action to generate UML flow diagrams from Java code
@@ -100,411 +99,20 @@ public class SequenceDiagramAction extends AnAction implements DumbAware {
         // 收集代码
         String collectedCode = collectCodeFromCallStack(callStack);
 
-        // 显示初始对话框，包含"生成流程"按钮
-        SwingUtilities.invokeLater(() ->
-            showInitialDialog(project, callStack, collectedCode, currentMethod.getClass().getSimpleName()+"."+currentMethod.getName()));
-    }
+        // 显示初始对话框
+        SwingUtilities.invokeLater(() -> {
+            String title = currentMethod.getClass().getSimpleName() + "." + currentMethod.getName();
 
-    /**
-     * 显示初始对话框，包含"生成流程"按钮
-     */
-    private void showInitialDialog(Project project, CallStack callStack, String collectedCode, String title) {
-        JFrame frame = new JFrame("UML时序图: " + title);
-        frame.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-        frame.setSize(1200, 800);
-
-
-        // 创建选项卡面板
-        JTabbedPane tabbedPane = new JTabbedPane();
-
-        // 创建初始PlantUML选项卡（空白）
-        JPanel plantUmlTab = createInitialPlantUmlTab(project, frame, collectedCode);
-        tabbedPane.addTab("PlantUML视图", plantUmlTab);
-
-
-        // 底部面板
-        JPanel bottomPanel = new JPanel(new FlowLayout(FlowLayout.RIGHT));
-        JLabel enhancedLabel = new JLabel();
-        enhancedLabel.setForeground(Color.BLUE);
-        bottomPanel.add(enhancedLabel);
-
-        // 创建主面板
-        JPanel mainPanel = new JPanel(new BorderLayout());
-        mainPanel.add(tabbedPane, BorderLayout.CENTER);
-        mainPanel.add(bottomPanel, BorderLayout.SOUTH);
-
-        UmlFlowService plugin = project.getService(UmlFlowService.class);
-        mainPanel.setName(title);
-        plugin.addFlow(mainPanel);
-
-//        frame.add(mainPanel);
-//        frame.setVisible(true);
-    }
-
-    /**
-     * 创建初始PlantUML选项卡，包含"生成流程"按钮
-     */
-    private JPanel createInitialPlantUmlTab(Project project, JFrame parentFrame, String collectedCode) {
-        JPanel panel = new JPanel(new BorderLayout());
-
-        // 创建一个分割面板，左侧显示代码，右侧显示图形
-        JSplitPane splitPane = new JSplitPane(JSplitPane.HORIZONTAL_SPLIT);
-        splitPane.setDividerLocation(500);
-        splitPane.setResizeWeight(0.5);
-
-        // 左侧代码面板
-        JPanel codePanel = new JPanel(new BorderLayout());
-        JTextArea textArea = new JTextArea();
-        textArea.setEditable(true);
-        textArea.setText(collectedCode);
-        textArea.setFont(new Font(Font.MONOSPACED, Font.PLAIN, 14));
-        codePanel.add(new JScrollPane(textArea), BorderLayout.CENTER);
-
-        // 右侧图形面板（初始为空白）
-        JPanel diagramPanel = new JPanel(new BorderLayout());
-        JLabel waitingLabel = new JLabel("点击\"生成流程\"按钮生成时序图", SwingConstants.CENTER);
-        waitingLabel.setFont(new Font(Font.SANS_SERIF, Font.BOLD, 16));
-        diagramPanel.add(waitingLabel, BorderLayout.CENTER);
-
-        // 添加到分割面板
-        splitPane.setLeftComponent(codePanel);
-        splitPane.setRightComponent(diagramPanel);
-
-        // 底部按钮面板
-        JPanel buttonPanel = new JPanel(new FlowLayout(FlowLayout.RIGHT));
-
-        // AI提供商选择下拉框
-        JLabel providerLabel = new JLabel("AI提供商:");
-        java.util.List<IdeaSettings.CustomAiProviderConfig> availableProviders = AiUtils.getCustomProviders();
-        JComboBox<IdeaSettings.CustomAiProviderConfig> providerComboBox = new JComboBox<>(availableProviders.toArray(new IdeaSettings.CustomAiProviderConfig[0]));
-
-        // 具体模型选择下拉框
-        JLabel specificModelLabel = new JLabel("具体模型:");
-        JComboBox<String> specificModelComboBox = new JComboBox<>();
-
-        // 监听提供商改变时，更新可用模型列表
-        providerComboBox.addActionListener(e -> {
-            IdeaSettings.CustomAiProviderConfig selected = (IdeaSettings.CustomAiProviderConfig) providerComboBox.getSelectedItem();
-            specificModelComboBox.removeAllItems();
-            if (selected != null && selected.getModels() != null) {
-                String[] models = selected.getModels().split(",");
-                for (String model : models) {
-                    specificModelComboBox.addItem(model.trim());
-                }
-                if (models.length > 0) {
-                    specificModelComboBox.setSelectedIndex(0);
-                }
-            }
-        });
-
-        // 设置默认选择
-        if (!availableProviders.isEmpty()) {
-            providerComboBox.setSelectedIndex(-1);
-            providerComboBox.setSelectedIndex(0);
-        }
-
-        providerComboBox.setRenderer(new DefaultListCellRenderer() {
-            @Override
-            public Component getListCellRendererComponent(JList<?> list, Object value, int index, boolean isSelected, boolean cellHasFocus) {
-                super.getListCellRendererComponent(list, value, index, isSelected, cellHasFocus);
-                if (value instanceof IdeaSettings.CustomAiProviderConfig) {
-                    IdeaSettings.CustomAiProviderConfig provider = (IdeaSettings.CustomAiProviderConfig) value;
-                    setText(provider.getName());
-                }
-                return this;
-            }
-        });
-
-        buttonPanel.add(providerLabel);
-        buttonPanel.add(providerComboBox);
-        buttonPanel.add(specificModelLabel);
-        buttonPanel.add(specificModelComboBox);
-
-        // 生成流程按钮
-        JButton generateButton = new JButton("生成流程");
-        generateButton.addActionListener(e -> {
-            // 禁用按钮，防止重复点击
-            generateButton.setEnabled(false);
-            generateButton.setText("生成中...");
-
-            // 获取选择的AI提供商
-            IdeaSettings.CustomAiProviderConfig selectedProvider = (IdeaSettings.CustomAiProviderConfig) providerComboBox.getSelectedItem();
-            String selectedModel = (String) specificModelComboBox.getSelectedItem();
-            if (selectedProvider == null) {
-                LOG.info("No AI provider selected or available");
-                Notifications.Bus.notify(new Notification(
-                        "com.yt.huq.idea",
-                        "无可用AI模型",
-                        "未选择AI提供商或没有可用的AI提供商。请在设置 > UmlFlowAiConfigurable 中配置至少一个提供商",
-                        NotificationType.ERROR),
-                        project);
-                // 重新启用按钮
-                generateButton.setEnabled(true);
-                generateButton.setText("生成流程");
-                return;
-            }
-            if (IdeaSettings.getInstance().getState().getPlantumlPathVal() == null ||
-                    IdeaSettings.getInstance().getState().getPlantumlPathVal().trim().isEmpty()) {
-                LOG.info("PlantUML安装路径未配置，无法渲染时序图");
-                Notifications.Bus.notify(new Notification(
-                        "com.yt.huq.idea",
-                        "PlantUML路径缺失",
-                        "PlantUML路径未配置。请在设置 > UmlFlowAiConfigurable 中设置",
-                        NotificationType.ERROR),
-                        project);
-                // 重新启用按钮
-                generateButton.setEnabled(true);
-                generateButton.setText("生成流程");
-                return;
-            }
-
-            // 在后台任务中执行AI调用
-            new Task.Backgroundable(project, "生成时序图", true, PerformInBackgroundOption.ALWAYS_BACKGROUND) {
+            UmlDiagramUIFactory.PromptProvider promptProvider = new UmlDiagramUIFactory.PromptProvider() {
                 @Override
-                public void run(@NotNull ProgressIndicator indicator) {
-                    indicator.setText("正在生成PlantUML时序图...");
-                    indicator.setIndeterminate(true);
-
-                    // 生成PlantUML时序图
+                public String getPrompt(String collectedCode) {
                     String flowPromptTemplate = getFlowDiagramPrompt();
-                    String flowPrompt = String.format(flowPromptTemplate, collectedCode);
-                    
-                    // 使用选择的AI模型调用
-                    AiUtils.AiConfig config = new AiUtils.AiConfig(selectedProvider, selectedModel);
-                    if (config.getApiKey() == null || config.getApiKey().trim().isEmpty()) {
-                        SwingUtilities.invokeLater(() -> {
-                            generateButton.setEnabled(true);
-                            Notifications.Bus.notify(new Notification(
-                                "com.yt.huq.idea",
-                                "API密钥未配置",
-                                "请在设置中为 " + selectedProvider.getName() + " 配置API密钥",
-                                NotificationType.WARNING),
-                                project);
-                        });
-                        return;
-                    }
-                    
-                    config.setSystemMessage("你是一个专业的PlantUML时序图生成专家，擅长分析Java代码并生成高质量的时序图。")
-                          .setTemperature(0.7)
-                          .setMaxTokens(8000);
-                    
-                    AiUtils.AiResponse response = AiUtils.callAi(flowPrompt, config);
-                    String flowDiagram = response.isSuccess() ? response.getContent() : null;
-
-
-                    if (flowDiagram != null && !flowDiagram.isEmpty()) {
-                        // 清理响应
-                        flowDiagram = cleanupUmlResponse(flowDiagram);
-
-                        // 更新UI
-                        String finalFlowDiagram = flowDiagram;
-                        SwingUtilities.invokeLater(() -> {
-                            // 更新文本区域
-                            textArea.setText(finalFlowDiagram);
-
-                            // 更新图形面板
-                            JPanel newDiagramPanel = PlantUmlRenderer.createPlantUmlPanel(finalFlowDiagram);
-                            splitPane.setRightComponent(newDiagramPanel);
-
-
-                            // 重新启用按钮
-                            generateButton.setEnabled(true);
-                            generateButton.setText("重新生成");
-
-                            // 刷新UI
-                            panel.revalidate();
-                            panel.repaint();
-                        });
-                    } else {
-                        SwingUtilities.invokeLater(() -> {
-                            String errorMsg = "生成时序图失败，请检查API设置和网络连接";
-                            if (response != null && !response.isSuccess() && response.getErrorMessage() != null) {
-                                errorMsg = "生成时序图失败: " + response.getErrorMessage();
-                            }
-                            
-                            Notifications.Bus.notify(new Notification(
-                                    "com.yt.huq.idea",
-                                    "时序图生成",
-                                    errorMsg,
-                                    NotificationType.ERROR),
-                                    project);
-
-                            // 重新启用按钮
-                            generateButton.setEnabled(true);
-                            generateButton.setText("生成流程");
-                        });
-                    }
+                    return String.format(flowPromptTemplate, collectedCode);
                 }
-            }.queue();
+            };
+
+            UmlDiagramUIFactory.showInitialDialog(project, collectedCode, "UML时序图: " + title, promptProvider, "生成时序图");
         });
-        buttonPanel.add(generateButton);
-
-        // 复制按钮
-        JButton copyButton = new JButton("复制到剪贴板");
-        copyButton.addActionListener(e -> {
-            copyToClipboard(textArea.getText());
-            Notifications.Bus.notify(new Notification(
-                    "com.yt.huq.idea",
-                    "UML图表",
-                    "UML图表已复制到剪贴板",
-                    NotificationType.INFORMATION),
-                    project);
-        });
-        buttonPanel.add(copyButton);
-
-        // 保存代码按钮
-        JButton saveCodeButton = new JButton("保存代码");
-        saveCodeButton.addActionListener(e -> {
-            JFileChooser fileChooser = new JFileChooser();
-            fileChooser.setDialogTitle("保存UML代码");
-            fileChooser.setSelectedFile(new java.io.File(parentFrame.getTitle().replaceAll("[^a-zA-Z0-9]", "_") + ".puml"));
-
-            int userSelection = fileChooser.showSaveDialog(parentFrame);
-            if (userSelection == JFileChooser.APPROVE_OPTION) {
-                java.io.File fileToSave = fileChooser.getSelectedFile();
-                try {
-                    java.io.FileWriter writer = new java.io.FileWriter(fileToSave);
-                    writer.write(textArea.getText());
-                    writer.close();
-                    Notifications.Bus.notify(new Notification(
-                            "com.yt.huq.idea",
-                            "UML图表",
-                            "UML代码已保存到 " + fileToSave.getAbsolutePath(),
-                            NotificationType.INFORMATION),
-                            project);
-                } catch (Exception ex) {
-                    Notifications.Bus.notify(new Notification(
-                            "com.yt.huq.idea",
-                            "UML图表",
-                            "保存UML代码失败: " + ex.getMessage(),
-                            NotificationType.ERROR),
-                            project);
-                }
-            }
-        });
-        buttonPanel.add(saveCodeButton);
-
-        // 保存图像按钮
-        JButton saveImageButton = new JButton("保存图像");
-        saveImageButton.addActionListener(e -> {
-            // 检查是否有内容可以保存
-            if (textArea.getText().trim().isEmpty() || !textArea.getText().contains("@startuml")) {
-                Notifications.Bus.notify(new Notification(
-                        "com.yt.huq.idea",
-                        "UML图表",
-                        "没有有效的UML图表可以保存",
-                        NotificationType.WARNING),
-                        project);
-                return;
-            }
-
-            JFileChooser fileChooser = new JFileChooser();
-            fileChooser.setDialogTitle("保存UML图像");
-            fileChooser.setSelectedFile(new java.io.File(parentFrame.getTitle().replaceAll("[^a-zA-Z0-9]", "_") + ".png"));
-
-            int userSelection = fileChooser.showSaveDialog(parentFrame);
-            if (userSelection == JFileChooser.APPROVE_OPTION) {
-                java.io.File fileToSave = fileChooser.getSelectedFile();
-                try {
-                    // 渲染并保存图像
-                    byte[] pngData = PlantUmlRenderer.renderPlantUmlToPng(textArea.getText());
-                    if (pngData != null) {
-                        java.io.FileOutputStream fos = new java.io.FileOutputStream(fileToSave);
-                        fos.write(pngData);
-                        fos.close();
-                        Notifications.Bus.notify(new Notification(
-                                "com.yt.huq.idea",
-                                "UML图表",
-                                "UML图像已保存到 " + fileToSave.getAbsolutePath(),
-                                NotificationType.INFORMATION),
-                                project);
-                    }
-                } catch (PlantUmlRenderException ex) {
-                    LOG.error("Failed to render UML diagram", ex);
-                    Notifications.Bus.notify(new Notification(
-                            "com.yt.huq.idea",
-                            "UML图表",
-                            "渲染UML图像失败: " + ex.getMessage(),
-                            NotificationType.ERROR),
-                            project);
-                } catch (Exception ex) {
-                    Notifications.Bus.notify(new Notification(
-                            "com.yt.huq.idea",
-                            "UML图表",
-                            "保存UML图像失败: " + ex.getMessage(),
-                            NotificationType.ERROR),
-                            project);
-                }
-            }
-        });
-        buttonPanel.add(saveImageButton);
-
-        // 刷新图像按钮
-        JButton refreshButton = new JButton("刷新图像");
-        refreshButton.addActionListener(e -> {
-            // 获取当前文本区域的内容
-            String updatedUmlContent = textArea.getText();
-
-            // 创建新的图形面板
-            JPanel newDiagramPanel = PlantUmlRenderer.createPlantUmlPanel(updatedUmlContent);
-
-            // 替换旧的图形面板
-            splitPane.setRightComponent(newDiagramPanel);
-            splitPane.revalidate();
-            splitPane.repaint();
-
-            // 刷新UI
-            panel.revalidate();
-            panel.repaint();
-        });
-        buttonPanel.add(refreshButton);
-
-        panel.add(splitPane, BorderLayout.CENTER);
-        panel.add(buttonPanel, BorderLayout.SOUTH);
-
-        return panel;
-    }
-
-
-    /**
-     * 复制文本到剪贴板
-     */
-    private void copyToClipboard(String text) {
-        StringSelection stringSelection = new StringSelection(text);
-        Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
-        clipboard.setContents(stringSelection, null);
-    }
-
-    /**
-     * 清理UML响应，确保它是有效的PlantUML代码
-     */
-    private String cleanupUmlResponse(String umlResponse) {
-        if (umlResponse == null || umlResponse.isEmpty()) {
-            return umlResponse;
-        }
-
-        // 确保以@startuml开始
-        if (!umlResponse.startsWith("@startuml")) {
-            int startIndex = umlResponse.indexOf("@startuml");
-            if (startIndex >= 0) {
-                umlResponse = umlResponse.substring(startIndex);
-            } else {
-                umlResponse = "@startuml\n" + umlResponse;
-            }
-        }
-
-        // 确保以@enduml结束
-        if (!umlResponse.endsWith("@enduml")) {
-            if (umlResponse.contains("@enduml")) {
-                // 截取到最后一个@enduml
-                int endIndex = umlResponse.lastIndexOf("@enduml") + "@enduml".length();
-                umlResponse = umlResponse.substring(0, endIndex);
-            } else {
-                umlResponse = umlResponse + "\n@enduml";
-            }
-        }
-
-        return umlResponse;
     }
